+++ packages/fcl/src/transaction/index.js
@@ -1,14 +1,21 @@
 import "../default-config"
 import {config} from "@onflow/config"
-import {spawn, send, INIT, SUBSCRIBE, UNSUBSCRIBE} from "@onflow/util-actor"
+import {
+  spawn,
+  send,
+  subscriber,
+  snapshoter,
+  UPDATED,
+  SNAPSHOT,
+  INIT,
+  SUBSCRIBE,
+  UNSUBSCRIBE,
+} from "@onflow/util-actor"
 import {send as fclSend} from "../send"
 import {decode} from "../decode"
 import {getTransactionStatus} from "@onflow/sdk"
 
-const POLL_RATE = 1000
-
-const UPDATED = "TRANSACTION/UPDATED"
-const SNAPSHOT = "SNAPSHOT"
+const RATE = 2500
 const POLL = "POLL"
 
 const fetchTxStatus = async transactionId => {
@@ -16,16 +23,21 @@ const fetchTxStatus = async transactionId => {
   return decode(response)
 }
 
-const isSealed = tx => tx.status === 4
-const isExecuted = tx => tx.status === 3 || isSealed(tx)
-const isFinalized = tx => tx.status === 2 || isExecuted(tx)
-const isPending = tx => tx.status === 1 || isFinalized(tx)
-const isUnknown = tx => tx.status === 0 || isPending(tx)
+const isExpired = tx => tx.status === 5
+const isSealed = tx => tx.status >= 4
+const isExecuted = tx => tx.status >= 3
+const isFinalized = tx => tx.status >= 2
+const isPending = tx => tx.status >= 1
+const isUnknown = tx => tx.status >= 0
+
+const isDiff = (cur, next) => {
+  return JSON.stringify(cur) !== JSON.stringify(next)
+}
 
 const HANDLERS = {
   [INIT]: async ctx => {
     const tx = await fetchTxStatus(ctx.self())
-    if (!isSealed(tx)) setTimeout(() => send(ctx.self(), POLL), POLL_RATE)
+    if (!isSealed(tx)) setTimeout(() => ctx.sendSelf(POLL), RATE)
     ctx.merge(tx)
   },
   [SUBSCRIBE]: (ctx, letter) => {
@@ -40,48 +52,42 @@ const HANDLERS = {
   },
   [POLL]: async ctx => {
     const tx = await fetchTxStatus(ctx.self())
-    if (!isSealed(tx)) setTimeout(() => send(ctx.self(), POLL), POLL_RATE)
+    if (!isSealed(tx)) setTimeout(() => ctx.sendSelf(POLL), RATE)
+    if (isDiff(ctx.all(), tx)) ctx.broadcast(UPDATED, tx)
     ctx.merge(tx)
-    ctx.broadcast(UPDATED, ctx.all())
   },
 }
 
-const spawnTransaction = transactionId => {
+const scoped = transactionId => {
   if (typeof transactionId === "object")
     transactionId = transactionId.transactionId
   if (transactionId == null) throw new Error("transactionId required")
-  return spawn(HANDLERS, transactionId)
+  return transactionId
+}
+
+const spawnTransaction = transactionId => {
+  return spawn(HANDLERS, scoped(transactionId))
 }
 
 export function transaction(transactionId) {
   function snapshot() {
-    const txId = spawnTransaction(transactionId)
-    return send(txId, SNAPSHOT, null, {expectReply: true, timeout: 10})
+    return snapshoter(transactionId, spawnTransaction)
   }
 
   function subscribe(callback) {
-    const txId = spawnTransaction(transactionId)
-    const EXIT = "@EXIT"
-    const self = spawn(async ctx => {
-      ctx.send(txId, SUBSCRIBE)
-      while (1) {
-        const letter = await ctx.receive()
-        if (letter.tag === EXIT) {
-          ctx.send(txId, UNSUBSCRIBE)
-          return
-        }
-        callback(letter.data)
-      }
-    })
-    return () => send(self, EXIT)
+    return subscriber(scoped(transactionId), spawnTransaction, callback)
   }
 
   function once(predicate) {
-    return function innerOnce() {
-      return new Promise(resolve => {
-        const unsub = subscribe(transaction => {
-          if (predicate(transaction)) {
-            resolve(transaction)
+    return function innerOnce(opts = {}) {
+      const suppress = opts.suppress || false
+      return new Promise((resolve, reject) => {
+        const unsub = subscribe(txStatus => {
+          if (txStatus.statusCode && !suppress) {
+            reject(txStatus.errorMessage)
+            unsub()
+          } else if (predicate(txStatus)) {
+            resolve(txStatus)
             unsub()
           }
         })
@@ -92,7 +98,6 @@ export function transaction(transactionId) {
   return {
     snapshot,
     subscribe,
-    oncePending: once(isPending),
     onceFinalized: once(isFinalized),
     onceExecuted: once(isExecuted),
     onceSealed: once(isSealed),
@@ -104,3 +109,4 @@ transaction.isPending = isPending
 transaction.isFinalized = isFinalized
 transaction.isExecuted = isExecuted
 transaction.isSealed = isSealed
+transaction.isExpired = isExpired