+++ packages/fcl/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 - YYYY-MM-DD **BREAKING?** -- description
 
+### 0.0.51 -- 2020-08-14
+
+- 2020-08-14 -- `fcl.tx(resp).once*` throw errors when statusCode is non-zero, errors can be suppressed (previous behaviour) py passing `{suppress: true}` to the `.once*` function.
+- 2020-08-14 -- `fcl.tx(resp).subscribe(callback)` callback will only be called when the status is different to what it was before.
+
 ### 0.0.50 -- 2020-08-14
 
 - 2020-08-14 -- Fixeds some small bugs due to typos